Move validate_measurements to conftest.
