+++ tests/test_compiler.py
@@ -18,10 +18,12 @@ Unit tests for the compiler subpackage.
 from unittest.mock import patch
 
 import pytest
+from conftest import validate_measurements
 
 import pennylane as qml
 from pennylane import numpy as np
 from pennylane.compiler.compiler import CompileError
+from pennylane.transforms.dynamic_one_shot import fill_in_value
 
 catalyst = pytest.importorskip("catalyst")
 jax = pytest.importorskip("jax")
@@ -737,3 +739,74 @@ class TestCatalystSample:
 
         assert circuit(0.0) == 0
         assert circuit(jnp.pi) == 1
+
+
+class TestCatalystMCMs:
+    """Test dynamic_one_shot with Catalyst."""
+
+    @pytest.mark.xfail(reason="requires simultaneous catalyst pr")
+    @pytest.mark.parametrize("measure_f", [qml.counts, qml.expval, qml.probs])
+    @pytest.mark.parametrize("meas_obj", [qml.PauliZ(0), [0], "mcm"])
+    # pylint: disable=too-many-arguments
+    def test_dynamic_one_shot_simple(self, measure_f, meas_obj):
+        """Tests that Catalyst yields the same results as PennyLane's DefaultQubit for a simple
+        circuit with a mid-circuit measurement."""
+        if measure_f in (qml.counts, qml.probs, qml.sample) and (
+            not isinstance(meas_obj, list) and not meas_obj == "mcm"
+        ):
+            pytest.skip("Can't use observables with counts, probs or sample")
+
+        if measure_f in (qml.var, qml.expval) and (isinstance(meas_obj, list)):
+            pytest.skip("Can't use wires/mcm lists with var or expval")
+
+        if measure_f == qml.var and (not isinstance(meas_obj, list) and not meas_obj == "mcm"):
+            pytest.xfail("isa<UnrealizedConversionCastOp>")
+        shots = 8000
+
+        dq = qml.device("default.qubit", shots=shots, seed=8237945)
+
+        @qml.defer_measurements
+        @qml.qnode(dq)
+        def ref_func(x, y):
+            qml.RX(x, wires=0)
+            m0 = qml.measure(0)
+            qml.cond(m0, qml.RY)(y, wires=1)
+
+            meas_key = "wires" if isinstance(meas_obj, list) else "op"
+            meas_value = m0 if isinstance(meas_obj, str) else meas_obj
+            kwargs = {meas_key: meas_value}
+            if measure_f == qml.counts:
+                kwargs["all_outcomes"] = True
+            return measure_f(**kwargs)
+
+        dev = qml.device("lightning.qubit", wires=2, shots=shots)
+
+        @qml.qjit
+        @catalyst.dynamic_one_shot
+        @qml.qnode(dev)
+        def func(x, y):
+            qml.RX(x, wires=0)
+            m0 = catalyst.measure(0)
+
+            @catalyst.cond(m0 == 1)
+            def ansatz():
+                qml.RY(y, wires=1)
+
+            ansatz()
+
+            meas_key = "wires" if isinstance(meas_obj, list) else "op"
+            meas_value = m0 if isinstance(meas_obj, str) else meas_obj
+            kwargs = {meas_key: meas_value}
+            return measure_f(**kwargs)
+
+        params = jnp.pi / 4 * jnp.ones(2)
+        results0 = ref_func(*params)
+        results1 = func(*params)
+        if measure_f == qml.counts and isinstance(meas_obj, list):
+            results1 = {
+                format(int(state), f"0{len(meas_obj)}b"): count for state, count in zip(*results1)
+            }
+        if measure_f == qml.sample:
+            results0 = results0[results0 != fill_in_value]
+            results1 = results1[results1 != fill_in_value]
+        validate_measurements(measure_f, shots, results1, results0)
